+++ test/test.js
@@ -1,19 +1,21 @@
 /*global describe, it */
 'use strict';
 
-var assert = require('assert');
-var fs = require('fs');
-var inline = require('../');
 var path = require('path');
+var fs = require('fs');
+
+var assert = require('assert');
 var rework = require('rework');
 
-describe('inline()', function () {
+var imprt = require('../');
+
+describe('imprt()', function () {
     it('should import stylsheets', function () {
         var original = fs.readFileSync(path.join(__dirname, 'fixtures/original.css'), 'utf8');
         var expected = fs.readFileSync(path.join(__dirname, 'fixtures/expected.css'), 'utf8');
 
         var css = rework(original)
-            .use(inline({ path: path.join(__dirname, 'fixtures') }))
+            .use(imprt({ path: path.join(__dirname, 'fixtures') }))
             .toString();
 
         assert.equal(css, expected);
@@ -24,7 +26,18 @@ describe('inline()', function () {
         var expected = fs.readFileSync(path.join(__dirname, 'fixtures/expected-recursive.css'), 'utf8');
 
         var css = rework(original)
-            .use(inline({ path: path.join(__dirname, 'fixtures') }))
+            .use(imprt({ path: path.join(__dirname, 'fixtures') }))
+            .toString();
+
+        assert.equal(css, expected);
+    });
+
+    it('should import stylsheets relatively', function () {
+        var original = fs.readFileSync(path.join(__dirname, 'fixtures/original-relative.css'), 'utf8');
+        var expected = fs.readFileSync(path.join(__dirname, 'fixtures/expected-relative.css'), 'utf8');
+
+        var css = rework(original)
+            .use(imprt({ path: path.join(__dirname, 'fixtures') }))
             .toString();
 
         assert.equal(css, expected);