+++ Changelog.md
@@ -6,6 +6,7 @@ Merge with [rework-inline](https://www.npmjs.org/package/rework-inline) bring th
 * Add encoding option
 * Use [find-file](https://www.npmjs.org/package/find-file) to lookup imports
 * Automated tests
+* relative `@import`
 
 # 0.0.3 / 2013-08-07